backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
pydantic>=2.6
pyyaml>=6.0
requests>=2.31
cryptography>=42.0
python-multipart>=0.0.9
pytest>=8.0
httpx>=0.27

# --- OSS storage backends (optional; required only when storage.backend != "local") ---
boto3>=1.34
oss2>=2.18
cos-python-sdk-v5>=1.9

